- changed type of httptest.lastfailedstep to integer (Alexei)
git-svn-id: svn://svn.zabbix.com/trunk@4136 f4e7fd02-d210-4e34-a1e0-d4093fb66cc2

+++ upgrades/dbpatches/1.4/mysql/patch/httptest.sql
@@ -6,7 +6,7 @@ CREATE TABLE httptest (
 	nextcheck		integer		DEFAULT '0'	NOT NULL,
 	curstate		integer		DEFAULT '0'	NOT NULL,
 	curstep		integer		DEFAULT '0'	NOT NULL,
-	lastfailedstep		bigint unsigned		DEFAULT '0'	NOT NULL,
+	lastfailedstep		integer	DEFAULT '0'	NOT NULL,
 	delay		integer		DEFAULT '60'	NOT NULL,
 	status		integer		DEFAULT '0'	NOT NULL,
 	macros		blob		DEFAULT ''	NOT NULL,
