+++ upgrades/dbpatches/1.4/postgresql/patch/httptest.sql
@@ -6,7 +6,7 @@ CREATE TABLE httptest (
 	nextcheck	integer		DEFAULT '0'	NOT NULL,
 	curstate	integer		DEFAULT '0'	NOT NULL,
 	curstep		integer		DEFAULT '0'	NOT NULL,
-	lastfailedstep	bigint DEFAULT '0'	NOT NULL,
+	lastfailedstep	integer		 DEFAULT '0'	NOT NULL,
 	delay		integer		DEFAULT '60'	NOT NULL,
 	status		integer		DEFAULT '0'	NOT NULL,
 	macros		text		DEFAULT ''	NOT NULL,
